Removed some debugging statements and fixed the CORBA type error.  Data still not flowing however...The input and output ports are showing the same number of elements per second, etc., however the output data is 0's.

+++ python/redhawk_source.py
@@ -53,7 +53,7 @@ class redhawk_source(gr.sync_block, ProvidesShort_i, OrbCreator):
                 self,
                 name="redhawk_source",
                 in_sig=None,
-                out_sig=[numpy.float])
+                out_sig=[numpy.short])
 
     def __del__(self):
         OrbCreator.__del__(self)
